+++ driver/src/test/unit/com/mongodb/reactivestreams/client/CustomMatchers.groovy
@@ -22,6 +22,7 @@ import org.hamcrest.Description
 @SuppressWarnings('NoDef')
 class CustomMatchers {
 
+    static nullList = [null, null]
     static isTheSameAs(final Object e) {
         [
                 matches         : { a -> compare(e, a) },
@@ -31,7 +32,20 @@ class CustomMatchers {
         ] as BaseMatcher
     }
 
+    static isTheSameAs(final Object e, final List<String> ignoreNames) {
+        [
+                matches         : { a -> compare(e, a, ignoreNames) },
+                describeTo      : { Description description -> description.appendText("Operation has the same attributes ${e.class.name}")
+                },
+                describeMismatch: { a, description -> describer(e, a, ignoreNames, description) }
+        ] as BaseMatcher
+    }
+
     static compare(expected, actual) {
+        compare(expected, actual, [])
+    }
+
+    static compare(expected, actual, ignoreNames) {
         if (expected == actual) {
             return true
         }
@@ -41,15 +55,17 @@ class CustomMatchers {
         if (actual.class.name != expected.class.name) {
             return false
         }
-        getFieldNames(actual.class).collect { it ->
+        getFieldNames(actual.class).findAll { !ignoreNames.contains(it) } .collect {
             if (nominallyTheSame(it)) {
                 return actual."$it".class == expected."$it".class
             } else if (actual."$it" != expected."$it") {
                 def (a1, e1) = [actual."$it", expected."$it"]
-                if (List.isCase(a1) && List.isCase(e1) && (a1.size() == e1.size())) {
+                if ([a1, e1].contains(null) && [a1, e1] != nullList) {
+                    return false
+                } else if (List.isCase(a1) && List.isCase(e1) && (a1.size() == e1.size())) {
                     def i = -1
                     return a1.collect { a -> i++; compare(a, e1[i]) }.every { it }
-                } else if (a1 != null && a1.class.name.startsWith('com.mongodb') && a1.class == e1.class) {
+                } else if (a1.class != null && a1.class.name.startsWith('com.mongodb') && a1.class == e1.class) {
                     return compare(a1, e1)
                 }
                 return false
@@ -59,6 +75,10 @@ class CustomMatchers {
     }
 
     static describer(expected, actual, description) {
+        describer(expected, actual, [], description)
+    }
+
+    static describer(expected, actual, ignoreNames, description) {
         if (expected == actual) {
             return true
         }
@@ -71,7 +91,7 @@ class CustomMatchers {
             return false
         }
 
-        getFieldNames(actual.class).collect { it ->
+        getFieldNames(actual.class).findAll { !ignoreNames.contains(it) } .collect {
             if (nominallyTheSame(it)) {
                 if (actual."$it".class != expected."$it".class) {
                     description.appendText("different classes $it :" +
@@ -80,15 +100,18 @@ class CustomMatchers {
                 }
             } else if (actual."$it" != expected."$it") {
                 def (a1, e1) = [actual."$it", expected."$it"]
-                if (List.isCase(a1) && List.isCase(e1) && (a1.size() == e1.size())) {
+                if (([a1, e1].contains(null) || [a1.class, e1.class].contains(null)) && [a1, e1] != nullList) {
+                    description.appendText("different values in $it : $e1 != $a1\n")
+                    return false
+                } else if (List.isCase(a1) && List.isCase(e1) && (a1.size() == e1.size())) {
                     def i = -1
                     a1.each { a ->
                         i++; if (!compare(a, e1[i])) {
                             describer(a, e1[i], description)
                         }
                     }.every { it }
-                } else if (a1 != null && a1.class.name.startsWith('com.mongodb') && a1.class == e1.class) {
-                    return compare(a1, e1, description)
+                } else if (a1.class.name.startsWith('com.mongodb') && a1.class == e1.class) {
+                    return describer(a1, e1, description)
                 }
                 description.appendText("different values in $it : $e1 != $a1\n")
                 return false
@@ -97,11 +120,11 @@ class CustomMatchers {
         }
     }
 
-    static getFieldNames(Class curClass) {
+    static List<String> getFieldNames(Class curClass) {
         getFieldNames(curClass, [])
     }
 
-    static getFieldNames(Class curClass, names) {
+    static List<String> getFieldNames(Class curClass, names) {
         if (curClass != Object) {
             getFieldNames(curClass.getSuperclass(), names += curClass.declaredFields.findAll { !it.synthetic }*.name)
         }